requirements.txt
torch>=2.0.0
torchvision>=0.15.0
clip @ git+https://github.com/openai/CLIP.git
qdrant-client>=1.1.1
numpy>=1.21.0  # Required by CLIP and torch
Pillow>=9.0.0  # Required by CLIP for image processing
